// data/github/current/ui/packages/react-sandbox/routes/__generated__/IssueSubIssues_NestedQuery.graphql.ts
export type IssueSubIssues_NestedQuery$variables = {
  id: string;
};
export type IssueSubIssues_NestedQuery$data = {
  readonly node: {
    readonly id?: string;
    readonly subIssues?: {
      readonly nodes: ReadonlyArray<{
        readonly " $fragmentSpreads": FragmentRefs<"IssueSubIssues_InnerFragment">;
      } | null | undefined> | null | undefined;
    };
  } | null | undefined;
};
export type IssueSubIssues_NestedQuery = {
  response: IssueSubIssues_NestedQuery$data;
  variables: IssueSubIssues_NestedQuery$variables;
};